This revision was mostly improvements to the sort menu. Now all torrents use data-* attributes, making the sort much more robust and faster (before it set a "sort" attribute for every time a sort was needed).
remote.php and rtorrent.php were cleaned up to make way for the new setsettings.php, which gives serverside rewrite functionality to prefs.json. This change has enabled Avalanche to retain settings such as if the details pane is open, and the method used to sort the torrents.

Also, this change brings about a license change to GPL v3, from the old MIT. Although more restrictive this should encourage developers to build upon our codebase and is the right decision in moving forward to a robust Avalanche. As part of this change the LICENSE.txt has been added.

This release also marks beta 4 or Avalanche - so all users should update.

git-svn-id: http://avalanche-rt.googlecode.com/svn/trunk@19 3051db82-1d7c-11df-8d00-ad93a377a9c4