total_diskspace as a setting
It will not work if downloading files are on more than 1 partition.
As rtorrent does not report total diskspace on the very downloading
torrent, with only rtorrent xmlrpc interface, we can do nothing to
change it.

An idea:
preset the total diskspace of all mount points that rtorrent are
saving files to them, eg:

{
  '/mnt/disk1': disksize1,
  '/mnt/disk2': disksize2,
}

than get the total diskspace according to where torrent files saved